GameManagerの改修 #12 (#15)
:move:GameManagerをメインプロジェクトに移動

---------
